requirements.txt
boto3==1.28.43
botocore==1.31.43
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==0.21.1
pandas==1.5.3
numpy==1.26.4
requests==2.32.2
